mods to earthlng patch #19
shortened and evened out lines, added that extra link. I changed "Internationalized Domain Names" to IDNs to save space and then realized the kb and wiki articles don;t even say what IDN stands for, so I put it back.

Also swapped the order and wording of the pref to make it consistent with the action. Instead of
- "2672: eliminate possible .. show_punycode", true)"
- "2672: force Punycode .. show_punycode", true)"