release policy update
"releases" which is the github term, is purposed for archiving legacy versions of the user.js. This is done *near* the end of each version's stable cycle (a week?), for the reasons given in the user.js. As soon as a "release" is done, the "live" version is incremented to the upcoming stable, and changes are started based on the diffs provided by earthlng.