0707: clean up DoH
- remove network.trr.bootstrapAddress = default blank anyway
- remove network.trr.uri - this is mozilla's default value. Unfortunately, that value is hard-coded in FF65+ into the UI settings as an option to use the default URI, so changing this could be misleading
- In FF65 there is now a custom URI field in the UI settings, and that's what users should be using - but it actually controls two prefs if you check to use custom - see https://github.com/ghacksuserjs/ghacks-user.js/issues/610#issuecomment-460169782

I am not interested in maintaining how this UI works (and it could change). The single pref left above is all that is needed to turns it on or off, that's all I care about. I've added the [SETTING] info and that's where people can do their shit. That's good enough for me