1820: update -> inactive, see #709
At best disabling the background update of gmp means not only an extra item for those who wish to use it (e.g widevine, netflix) to have to deal with, but also a time delay in getting the actual download. At worst, it could cause users to use an old dll (security risk).

I will leave it in, for now, but am seriously considering removing it, so don't cry if I do.