Update user.js
It's not deprecated.

Either it's an RFP alternative now that RFP indirectly covers it: i.e RFP prompts you on non en-US builds to spoof your language, and if you choose to spoof as en-US, then this solves that. If you're not spoofing your locale, then the pref has no significance.

Or we can remove it entirely: as I see no point in it TBH. If you're on en-US, then you're not leaking. If you use RFP + opt in to language spoofing, you're covered now. If you use RFP and opt out of language spoofing, then you're not leaking anything because you're telling the truth as per other language/locale metrics: i.e you don't need to spoof.

95% of me says we just remove the pref. 5% of me says - see what Earthlng thinks