0720: add network.trr.mode
This is a general rant (not aimed at anyone) .. enjoy

Getting fucking pressured and attacked at ghacks.net comments to make it active (both on and off), getting pressured here to fucking include (and fucking maintain it: yeah it's all good when you don't have to do the work), getting fucking criticized before I even do anything about it, getting fucking criticized for removing it, getting fucking criticized for not adding more to it (e.g bugs like trr=3) - and the whole fucking thing is a bit divisive and opinions on it are like assholes, every one has one. Whatever I do I am on a fucking losing streak and people will bitch and/or be sarcastic.

And this pref on it's own is fucking useless: it doesn't change anything as far as unwanted changes go. And I'm not fucking interested in more prefs (too many bugs, too many upcoming changes): one day, when some of the more serious bugs are fixed, and the UI is polished, then *maybe* the section 0720's can be expanded to include other settings. But by then, I can just say: go use the fucking UI.

Have I used the word fuck enough .. only eleven times .. should be more. Since 95% of the work falls on earthlng and myself: I reserve the right to not include a DoH section (for my own sanity)