prefsCleaner.bat: add -unattended flag
Usage:
  prefsCleaner.bat -unattended

Skips the prompt for user input and proceeds when -unattended is specified. If omitted, default behaviour is unchanged.

Code taken from updater.bat for consistency.

Please review this PR closely as I'm not fluent in batch scripting. The revised script appears to function properly with and without -unattended, but I'm not aware of common pitfalls.

Signed-off-by: Keith Harrison <keithh@protonmail.com>