updater.sh: POSIX friendly printing
echo flags aren't defined by POSIX. In general, anything more advanced
than a simple string followed by a newline should be handled by
printf(1) if portability/consistency are concerns.

Added new functions to handle this so that colors are kept separate from
the strings themselves (should make the script more maintainable in the
long run).

An additional benefit of using functions here is that errors and
warnings are more likely to be sent to STDERR.