updater.sh: adjust printf_color_newline()
printf_color() functions by assigning the first positional parameter and
shifting it out of the list of parameters. If the interface is misused
and the first positional parameter is also the only parameter, what
would happen? It is probably better to change both to handle that
situation the same way since the code savings is minimal.

While here, add back in a newline that was accidentally nuked in the
first commit.