Fail curl request on HTTP error status
Use the `-f' option to instruct curl to exit with a non-zero exit code
if the HTTP response indicates an error via its status code.  This is
the default behavior for wget, but curl will send the request and exit
normally (with exit code equals to zero), as long as it's received a
response from the server.