Add `-f' option to updater.sh for local user.js
Add the option `-f' to the updater.sh script to allow users to specify
a path in local disk to an up-to-date user.js file.  If specified, this
option overrides the default behavior of downloading the latest user.js
file.

This option was created due to GitHub blocking file downloads[1].

[1]: https://github.com/orgs/community/discussions/159123